requirements.txt
Django
djangorestframework
mysqlclient
python-dotenv
graphene_django
python-dateutil